extractGeometryInfo
